+++ ExportHDR.lrplugin/Info.lua
@@ -8,7 +8,7 @@ return {
 	LrSdkMinimumVersion = 14.0,
 	LrToolkitIdentifier = "com.karachungen.lightroom.export.ultrahdr",
 	LrPluginName = "Ultra HDR Export",
-	VERSION = { major = 2, minor = 0, revision = 4, build = 0 },
+	VERSION = { major = 2, minor = 0, revision = 5, build = 0 },
 
 	LrExportFilterProvider = {
 		{